chore(test): reorganize unit tests to match project structure + cleanup
