+++ test/jit/binding/unicode.ts
