+++ test/runtime/dom.spec.ts
